==> 1. Join Months Reddit Datasets for Trainining-FirstHalf.ipynb <==
import pandas as pd

# Reddit NLP Research - Posts 

dir_gen_posts      = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Months\01\MatchedPosts.csv'
dir_feb_posts      = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Months\02\MatchedPosts.csv'
dir_mar_posts      = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Months\03\MatchedPosts.csv'
dir_apr_posts      = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Months\04\MatchedPosts.csv'
dir_mag_posts      = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Months\05\MatchedPosts.csv'
dir_giu_posts      = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Months\06\MatchedPosts.csv'

### Gennaio

gen_posts = pd.read_csv(dir_gen_posts)
gen_posts = gen_posts.drop(columns=['Unnamed: 0'])
print(len(gen_posts))
gen_posts.head()

gen_posts = gen_posts[['full_text', 'keyword', 'keyword type']]

### Febbraio

feb_posts = pd.read_csv(dir_feb_posts)
feb_posts = feb_posts.drop(columns=['Unnamed: 0'])
print(len(feb_posts))
feb_posts.head()

### Marzo

mar_posts = pd.read_csv(dir_mar_posts)
mar_posts = mar_posts.drop(columns=['Unnamed: 0'])
print(len(mar_posts))
mar_posts.head()

### Aprile

apr_posts = pd.read_csv(dir_apr_posts)
apr_posts = apr_posts.drop(columns=['Unnamed: 0'])
print(len(apr_posts))
apr_posts.head()

### Maggio

mag_posts = pd.read_csv(dir_mag_posts)
mag_posts = mag_posts.drop(columns=['Unnamed: 0'])
print(len(mag_posts))
mag_posts.head()

### Giugno

giu_posts = pd.read_csv(dir_giu_posts)
giu_posts = giu_posts.drop(columns=['Unnamed: 0'])
print(len(giu_posts))
giu_posts.head()

# Join

df = pd.concat([gen_posts, feb_posts, mar_posts, apr_posts, mag_posts, giu_posts])
print(len(df))
df.head()

df['keyword type'].unique()

file_files = r'C:\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\ER_Schema\\Entries.csv'
behaviour_files = r'C:\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\ER_Schema\\Behaviours.csv'

df_files = pd.read_csv(file_files, low_memory = False)
df_files

all_malware_list = list(df_files.malware.unique())
all_malware_list = [item for item in all_malware_list if (item != 'unknown') & (item != 'malware_tag')]
all_malware_list

#df[df['keyword'].isna()]  # Replace NaN with an empty string
df[df['keyword type'] == 'malware']

for index, row in df.iterrows():
    if any(keyword in row['keyword'] for keyword in all_malware_list):
        df.at[index, 'keyword type'] = 'malware'
        
df

behaviours = pd.read_csv(behaviour_files)
behaviours.head()

all_behaviours_name_list = list(behaviours.action_name.unique())
all_behaviours_name_list = [ item for item in all_behaviours_name_list if (item != 'unknown') ]
all_behaviours_name_list

for index, row in df.iterrows():
    if any(keyword in row['keyword'] for keyword in all_behaviours_name_list):
        df.at[index, 'keyword type'] = 'behaviour'
        
df

# Save it 

df = df.sample(frac=1).reset_index(drop=True)
df

length = len(df)

split1 = int(0.25 * length)
split2 = int(0.5 * length)
split3 = int(0.75 * length)

df1 = df.iloc[:split1]
df2 = df.iloc[split1:split2]
df3 = df.iloc[split2:split3]
df4 = df.iloc[split3:]

df1.to_csv(r"C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Posts\df1.csv")
df2.to_csv(r"C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Posts\df2.csv")
df3.to_csv(r"C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Posts\df3.csv")
df4.to_csv(r"C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Reddit\Posts\df4.csv")
