==> 01.1. Data Collection - Threat Fox.ipynb <==
## 1.1. Data Collection source Threat Fox

import datetime
import os
from OTXv2 import OTXv2
import time
import json 
import pandas as pd 
import requests
import urllib3
import json
import sys

pd.options.display.max_colwidth = 500
pd.set_option('display.max_columns', None)

apiKey = 'a4582c8cb2e30fc6b444b509bb6a9a56'
day = 7

headers = { 'API-KEY': apiKey }

data = {
    'query': 'get_iocs',
    'days' : day
}

pool = urllib3.HTTPSConnectionPool('threatfox-api.abuse.ch', port=443, maxsize=500)

json_data = json.dumps(data)
response = pool.request("POST", "/api/v1/", body=json_data)
response = response.data.decode("utf-8", "ignore")
json_response =  json.loads(response)


entries = pd.DataFrame.from_dict(pd.json_normalize(json_response))
df = pd.DataFrame.from_dict(pd.json_normalize((pd.DataFrame.from_dict(pd.json_normalize(entries['data'])).T)[0]))
df

df.columns

len(df)

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\Threat_Fox.csv')
