==> 01.2. Data Collection - URL Haus.ipynb <==
## 1.2. Data Collection source URL Haus

import csv, urllib.request
import pandas as pd 
import requests

pd.options.display.max_colwidth = 500
pd.set_option('display.max_columns', None)

json_URL = 'https://urlhaus.abuse.ch/downloads/csv_recent/'

with requests.Session() as s:
    download = s.get(json_URL)

    df = pd.read_csv(json_URL, header = 8,on_bad_lines='skip')
    
df

df.columns

len(df)

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\URL_HAUS.csv')
