==> 01.3. Data Collection - Feodo Tracker.ipynb <==
## 1.3. Data Collection source Feodo Tracker

import datetime
import os
from OTXv2 import OTXv2
import time
import json 
import pandas as pd 
import requests
import urllib3
import json
import sys

pd.options.display.max_colwidth = 500
pd.set_option('display.max_columns', None)

import csv
import requests

json_URL = 'https://feodotracker.abuse.ch/downloads/ipblocklist.json'


with requests.Session() as s:
    download = s.get(json_URL)

    decoded_content = download.content.decode('utf-8','ignore')
    decoded_cleaned_content = decoded_content.replace('\n','')
    data = json.loads(decoded_cleaned_content)
    df = pd.DataFrame.from_dict(pd.json_normalize(data))
    
df

df.columns

len(df)

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\Feodo_Tracker.csv')
