==> 02.1. ER Schema Creation - Theat Fox .ipynb <==
## 2.1. ER Schema Creation - Threat Fox 

## Threat Fox Entities 

- Entries
- Tags
- Malwares
- Alias


import datetime
import os
from OTXv2 import OTXv2
import time
import json 
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 500

file_TF = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataCollection\Threat_Fox.csv'

# THREAT FOX

df = pd.read_csv(file_TF, low_memory = False )
df = df.drop('Unnamed: 0', axis=1)
df.head(30)

df.columns

### Column Filtering

df = df.drop([ 'threat_type_desc', 'ioc_type_desc', 'malware', 'confidence_level'], axis=1)
df

### Tags - Threat Fox

- TAGS ( ENTRY_ID, SOURCE , tag_name )

def split_list(row):
    row["tags"] = str(row["tags"]).replace('[','').replace(']','').replace("'",'').split(",")
    #print(row["tags"])
    return pd.Series(row['tags'])

tags = pd.DataFrame(df[['id','tags']])
tags

# apply the function to the DataFrame
new_df = tags.apply(split_list, axis=1).rename(columns=lambda x: f"tags{x+1}")
# merge the new DataFrame with the original DataFrame
tags = pd.concat([tags, new_df], axis=1).drop('tags', axis=1)
tags = tags.rename(columns={"id": "ID_ENTRY"})


columns = list(tags.columns)

new_df = pd.DataFrame()

for i, col in enumerate(columns):
    tmp_df = pd.DataFrame()
    if col != 'ID_ENTRY':
        
        tmp_df = tags[['ID_ENTRY', col]].rename(columns = {col : 'tag'})
        new_df = pd.concat([new_df, tmp_df], axis = 0)

Threat_Fox_TAGS_df = new_df

Threat_Fox_TAGS_df['source'] = 'Threat Fox'
Threat_Fox_TAGS_df = Threat_Fox_TAGS_df.sort_values(by='ID_ENTRY')
Threat_Fox_TAGS_df = Threat_Fox_TAGS_df.dropna(subset=['tag'], axis=0)
Threat_Fox_TAGS_df = Threat_Fox_TAGS_df.reset_index(drop = True)
Threat_Fox_TAGS_df = Threat_Fox_TAGS_df[['ID_ENTRY', 'source', 'tag']]
Threat_Fox_TAGS_df

### ALIAS - Threat Fox

- ALIAS ( MALWARE , ALIAS )


def split_list_malware(row):
    row["malware_alias"] = str(row["malware_alias"]).split(",")
    return pd.Series(row['malware_alias'])

alias = pd.DataFrame(df[['malware_printable','malware_alias']])

# apply the function to the DataFrame
new_df = alias.apply(split_list_malware, axis=1).rename(columns=lambda x: f"alias{x+1}")
alias = pd.concat([alias, new_df], axis=1).drop('malware_alias', axis=1)
alias = alias.rename(columns={"malware_printable": "malware"})

columns = list(alias.columns)

new_df = pd.DataFrame()

for i, col in enumerate(columns):
    tmp_df = pd.DataFrame()
    if col != 'malware':
        
        tmp_df = alias[['malware', col]].rename(columns = {col : 'alias'})
        new_df = pd.concat([new_df, tmp_df], axis = 0, ignore_index= True)
        
Threat_Fox_ALIAS_df = new_df

Threat_Fox_ALIAS_df = Threat_Fox_ALIAS_df.reset_index(drop = True)

Threat_Fox_ALIAS_df

### MALWARES - Threat Fox

- MALWARES ( MALWARE_NAME )


Threat_Fox_MALWARES_df = pd.DataFrame(Threat_Fox_ALIAS_df['malware']).drop_duplicates()
        
Threat_Fox_MALWARES_df = Threat_Fox_MALWARES_df.reset_index(drop = True)

Threat_Fox_MALWARES_df['type'] = 'unknown'
Threat_Fox_MALWARES_df

### ENTRIES - Threat Fox

- ENTRIES ( // ID_ENTRY, SOURCE // , ioc, ioc_type, threat_type, malware, first_seen, last_seen, reporter, reference )

columns_ordered = ['ID_ENTRY', 'source', 'ioc', 'ioc_type', 'threat_type', 'malware', 'first_seen', 'last_seen', 'reporter', 'reference']

Threat_Fox_ENTRIES_df = df.drop(columns=['malware_alias', 'tags'])

Threat_Fox_ENTRIES_df

# SAVE into ..\PostDataStructurization\Threat_Fox\ folder

dataframes_dict = {}

# Add the DataFrames to the dictionary with keys
dataframes_dict['Entries']  = Threat_Fox_ENTRIES_df
dataframes_dict['Tags']     = Threat_Fox_TAGS_df
dataframes_dict['Malwares'] = Threat_Fox_MALWARES_df
dataframes_dict['Alias']    = Threat_Fox_ALIAS_df

save_directory = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\"
for key, df in dataframes_dict.items():
    file_name = f"{key}.csv"
    file_path = save_directory + file_name
    df.to_csv(file_path, index=False)
