==> 02.2. ER Schema Creation - URL Haus.ipynb <==
## 2.2. ER Schema Creation - URL Haus 

## URL HAUS - Entities 

- Entries
- Tags


import datetime
import os
from OTXv2 import OTXv2
import time
import json 
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 500

file_UH = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataCollection\URL_Haus.csv'

# URL HAUS

df = pd.read_csv(file_UH, low_memory = False )
df = df.drop('Unnamed: 0', axis=1)
df.head(30)

df = df.rename(columns={'# id' : 'ID_ENTRY', 'url' : 'ioc', 'url_status' : 'status', 'dateadded' :'first_seen', 'last_online' :'last_seen', 'urlhaus_link' :'reference', 'threat' :'threat_type' })
df

### Entries - URL Haus

- Entries ( ENTRY_ID, SOURCE , ioc, threat_type, first_seen, last_seen, status, reporter, reference )

df['source'] = 'URL Haus'

ordered_columns = ['ID_ENTRY', 'source', 'ioc', 'threat_type', 'first_seen', 'last_seen', 'status', 'reporter', 'reference']
URL_HAUS_Entries_df = df[ordered_columns]
URL_HAUS_Entries_df

### Tags - URL Haus

- TAGS ( ENTRY_ID, SOURCE , tag_name )

ordered_tags_columns = ['ID_ENTRY', 'source', 'tags']
URL_HAUS_Tags_df = df[ordered_tags_columns]
URL_HAUS_Tags_df = URL_HAUS_Tags_df.rename(columns={'tags' : 'tag' })

URL_HAUS_Tags_df

# SAVE into ..\PostDataStructurization\URL_Haus\ folder

dataframes_dict = {}

# Add the DataFrames to the dictionary with keys
dataframes_dict['Entries']  = URL_HAUS_Entries_df
dataframes_dict['Tags']     = URL_HAUS_Tags_df

save_directory = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\URL_Haus\\"
for key, df in dataframes_dict.items():
    file_name = f"{key}.csv"
    file_path = save_directory + file_name
    df.to_csv(file_path, index=False)
