==> 02.3. ER Schema Creation - Feodo Tracker.ipynb <==
## 2.3. ER Schema Creation - Feodo Tracker

## Feodo Tracker Entities 

- Entries
- Tags
- Malwares
- Alias


import datetime
import os
from OTXv2 import OTXv2
import time
import json 
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 500

file_FT = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataCollection\Feodo_Tracker.csv'

# THREAT FOX

df = pd.read_csv(file_FT, low_memory = False )
df = df.drop('Unnamed: 0', axis=1)
df.head(30)

df.columns

## Merge columns
Merge column ip-address and port into a single column


df['ip'] = df['ip_address'].astype(str).str.cat(df['port'].astype(str), sep=':')
df = df.drop(['ip_address'], axis=1)
df = df.drop(['port'], axis=1)
df

### Column Filtering

df = df.drop([ 'hostname', 'as_number', 'as_name', 'country'], axis=1)
df

### Column Renaming

Feodo_Tracker_ENTRIES_df = df.rename(columns={ "last_online": "last_seen", "ip" : "ioc" })

Feodo_Tracker_ENTRIES_df["ioc_type"] = "ip:port"
Feodo_Tracker_ENTRIES_df["source"] = "Feodo Tracker"

Feodo_Tracker_ENTRIES_df['threat_type'] = 'botnet'
Feodo_Tracker_ENTRIES_df['reference'] = 'https://feodotracker.abuse.ch/'
Feodo_Tracker_ENTRIES_df['reporter'] = 'Feodo Tracker blocklist'

Feodo_Tracker_ENTRIES_df['ID_ENTRY'] = Feodo_Tracker_ENTRIES_df.index
Feodo_Tracker_ENTRIES_df

### MALWARES - Feodo Tracker

- MALWARES ( MALWARE_NAME )


Feodo_Tracker_MALWARES_df = pd.DataFrame(df['malware']).drop_duplicates()
        
Feodo_Tracker_MALWARES_df = Feodo_Tracker_MALWARES_df.reset_index(drop = True)

Feodo_Tracker_MALWARES_df['type'] = 'unknown'
Feodo_Tracker_MALWARES_df

### ENTRIES - Threat Fox

- ENTRIES ( // ID_ENTRY, SOURCE // , ioc, ioc_type, threat_type, malware, first_seen, last_seen, reporter, reference )

columns_ordered = ['ID_ENTRY', 'source', 'ioc', 'ioc_type', 'threat_type', 'malware', 'first_seen', 'last_seen', 'reporter', 'reference']

Feodo_Tracker_ENTRIES_df = Feodo_Tracker_ENTRIES_df[columns_ordered]

Feodo_Tracker_ENTRIES_df

# SAVE into ..\PostDataStructurization\Threat_Fox\ folder

dataframes_dict = {}

# Add the DataFrames to the dictionary with keys
dataframes_dict['Entries']  = Feodo_Tracker_ENTRIES_df
dataframes_dict['Malwares'] = Feodo_Tracker_MALWARES_df

save_directory = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Feodo_Tracker\\"
for key, df in dataframes_dict.items():
    file_name = f"{key}.csv"
    file_path = save_directory + file_name
    df.to_csv(file_path, index=False)
