==> 03.1. Data Profiling - Threat Fox.ipynb <==
# 3.1. Data Profiling - Threat Fox

import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

entries  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Entries.csv"
tags     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Tags.csv"
malwares = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Malwares.csv"
alias    = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Alias.csv"

## Data Profiling - Entries

entries_df = pd.read_csv(entries, low_memory = False )
entries_df

print("Duplicates: ",entries_df.duplicated().any())
#total number of not null values
not_null_values = entries_df.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = entries_df.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = entries_df.shape[0]*entries_df.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN']
ds_property_nan = pd.read_csv(entries, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

## Completeness

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

## Timeliness

from datetime import datetime

temp_columns = ['first_seen']

dateparse = lambda x: datetime.strptime(str(x), '%Y-%m-%d %H:%M:%S UTC')
ds_property_t = pd.read_csv(entries, parse_dates= temp_columns, date_parser=dateparse)

volatility = 20
today = datetime(2023, 9, 15)
ds_property_t['currency'] = (today-ds_property_t['first_seen']).dt.days

#adding a column with the timeliness computation
#if volatility is greater than currency the timeliness is equal to 1-currency/volatility, otherwise is 0
ds_property_t['Timeliness'] = np.where(volatility>ds_property_t['currency'],
                                       1-(ds_property_t['currency']/volatility),
                                       0)

print("Average Timeliness: ", ds_property_t['Timeliness'].mean())
print("Maximum Timeliness:", ds_property_t['Timeliness'].max())
print("Minimum Timeliness:", ds_property_t['Timeliness'].min())

### Profiling Auto-report

pip install pandas-profiling

pip install --upgrade numba

pip install visions==0.7.4

pip install --upgrade pydantic pydantic-settings

pip install ydata-profiling

from ydata_profiling import ProfileReport

profile = ProfileReport(entries_df, title="Threat Fox - Entries - Pandas Profiling Report")

profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Entries.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Entries.json")

## Value Distribution

#### Distribution of malwares

#import libraries and set up
import seaborn as sns
import matplotlib
matplotlib.use('tkagg')

import matplotlib.pyplot as plt


sns.set(color_codes=True)
sns.set_palette(sns.color_palette("muted"))

entries_df['first_seen'] = pd.to_datetime(entries_df['first_seen'], utc=True)


def distibution_plot(df, field):
    
    unique_malwares = df[field].unique()
    
    new = pd.DataFrame()
    
    s = 0
    
    nunique = len(unique_malwares)
    
    for i in range(nunique):
        
        malware = unique_malwares[i]
        
        s = s + len(df[df[field] == malware ])
    
    avg = ( s / nunique )
    
    for i in range(nunique):
        
        malware = unique_malwares[i]
        
        tmp = df[df[field] == malware ]
    
        if(len(tmp) > avg):
            
            new = pd.concat([new, tmp])       
            
    return new       

distibution_plot(entries_df, 'malware_printable')['malware_printable'].hist()


#### Distribution entries over time

# Create a time series plot with different colors for each category
plt.figure(figsize=(10, 6))

# Plot Category A with a blue color
plt.plot(entries_df['first_seen'], label='Threat Fox', marker='o', linestyle='-', color='blue')

# Customize the plot
plt.title('Time Series Data Threat Fox')
plt.xlabel('Timestamp')
plt.grid(True)

# Add a legend to distinguish between categories
plt.legend()

# Show the plot
plt.tight_layout()
plt.show()

## Data Profiling - Malwares

malwares_df = pd.read_csv(malwares, low_memory = False )
malwares_df

print("Duplicates: ",malwares_df.duplicated().any())
#total number of not null values
not_null_values = malwares_df.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = malwares_df.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = malwares_df.shape[0]*malwares_df.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN']
ds_property_nan = pd.read_csv(malwares, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

## Completeness

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

### Profiling Auto-report

profile = ProfileReport(entries_df, title="Threat Fox - Malwares - Pandas Profiling Report")

profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Malwares.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Malwares.json")

## Data Profiling - Alias

alias_df = pd.read_csv(alias, low_memory = False )
alias_df

print("Duplicates: ",alias_df.duplicated().any())
#total number of not null values
not_null_values = alias_df.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = alias_df.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = alias_df.shape[0]*alias_df.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN']
ds_property_nan = pd.read_csv(alias, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

## Completeness

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

### Profiling Auto-report

profile = ProfileReport(entries_df, title="Threat Fox - Alias - Pandas Profiling Report")

profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Alias.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Alias.json")

## Data Profiling - Tags

tags_df = pd.read_csv(tags, low_memory = False )
tags_df

print("Duplicates: ",tags_df.duplicated().any())
#total number of not null values
not_null_values = tags_df.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = tags_df.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = tags_df.shape[0]*alias_df.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN']
ds_property_nan = pd.read_csv(tags, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

## Completeness

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

### Profiling Auto-report

profile = ProfileReport(entries_df, title="Threat Fox - Tags - Pandas Profiling Report")

profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Tags.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 3\\Threat_Fox\\Tags.json")
