==> 04.1. Data Cleaning - Threat Fox.ipynb <==
# 4.1. Data Cleaning - source Threat Fox

import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

entries  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Entries.csv"
tags     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Tags.csv"
malwares = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Malwares.csv"
alias    = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataStructurization\\Threat_Fox\\Alias.csv"

## Entries - Threat Fox
### Data Cleaning

#### Error Correction - Missing Values

Drop columns affected by high missings :
- reference

entries_df = pd.read_csv(entries, low_memory = False )
entries_df

entries_df = entries_df.drop([ 'reference'], axis=1)
entries_df

duplicati = entries_df[entries_df.duplicated(subset=['ioc'], keep=False)]
duplicati

entries_df[entries_df.ioc == '45.61.138.12:443']

Imputation: for 'last_seen' column we could use imputation related to the 'first_seen' column, based on the implicit dependency. I assign the first_seen value if it's missing

entries_df['last_seen'].where(~(entries_df.last_seen.isnull()), other=entries_df.first_seen, inplace=True) 
entries_df

Imputation: if the field malware_printable is written 'Unknown malware' is set to np.NaN

entries_df['malware_printable'].where(~(entries_df.malware_printable == 'Unknown malware'), other= np.NaN, inplace=True) 

#### Standardization
in the field 'threat_type' we replace botnet_cc into botnet

entries_df['threat_type'].where(~(entries_df.threat_type == 'botnet_cc'), other='botnet', inplace=True) 

- ENTRIES ( // ID_ENTRY, SOURCE // , ioc, ioc_type, threat_type, malware, first_seen, last_seen, reporter, reference )

columns_ordered = ['ID_ENTRY', 'source', 'ioc', 'ioc_type', 'threat_type', 'malware', 'first_seen', 'last_seen', 'reporter', 'reference']
Threat_Fox_ENTRIES_df = entries_df.rename(columns={"id": "ID_ENTRY", "malware_malpedia" : "reference" , 'malware_printable' : 'malware'})
Threat_Fox_ENTRIES_df["source"] = "Threat Fox"
Threat_Fox_ENTRIES_df = Threat_Fox_ENTRIES_df[columns_ordered]

Threat_Fox_ENTRIES_df

### Normalization
All the chars of malwares are lower cases and all the punctualization factors are deleted

import math

def prepare_for_similarity_comparison(df, col_i, col_f):
    
    for i in range(len(df)):
        
        value = df.at[i,col_i]

        if (not isinstance(value, float) or not math.isnan(value)):    
            
            value = ' ' + value + ' ' 
            value = value.lower() 
            value = value.replace('?', ' ').replace('!', ' ').replace('%', ' ').replace('.', ' ').replace('-', ' ').replace('[' , ' ').replace( ']', ' ').replace('(' , ' ').replace( ')', ' ').replace('\\',' ').replace( ',', ' ').replace( '_', ' ') 
            value = value.strip() 
        
        df.loc[i, col_f] = value
            
    return df

Threat_Fox_ENTRIES_df = prepare_for_similarity_comparison(Threat_Fox_ENTRIES_df, 'malware', 'malware')
Threat_Fox_ENTRIES_df

## Tags - Threat Fox
### Data Cleaning

#### Error Correction

All the 3 attributes are keys of the entity, so if one of the them is Nan, the row is dropped

tags_df = pd.read_csv(tags, low_memory = False )
tags_df

Threat_Fox_TAGS_df = tags_df.dropna().reset_index(drop = True)

#### Data Normalization 

All the chars of tag are lower cases and all the punctualization factors are deleted

Threat_Fox_TAGS_df = prepare_for_similarity_comparison(Threat_Fox_TAGS_df, 'tag', 'tag')
Threat_Fox_TAGS_df

#### Data Deduplications 

Drop duplicates

Threat_Fox_TAGS_df = Threat_Fox_TAGS_df.drop_duplicates().reset_index(drop = True)
Threat_Fox_TAGS_df

## Malwares - Threat Fox
### Data Cleaning

#### Error Correction


malwares_df = pd.read_csv(malwares, low_memory = False )
malwares_df

Threat_Fox_MALWARES_df = malwares_df.dropna().reset_index(drop = True)

#### Data Normalization 

All the chars of tag are lower cases and all the punctualization factors are deleted

Threat_Fox_MALWARES_df = prepare_for_similarity_comparison(Threat_Fox_MALWARES_df, 'malware', 'malware')
Threat_Fox_MALWARES_df

#### Data Deduplications 

Drop duplicates

Threat_Fox_MALWARES_df = Threat_Fox_MALWARES_df.drop_duplicates()
Threat_Fox_MALWARES_df = Threat_Fox_MALWARES_df.reset_index(drop = True)
Threat_Fox_MALWARES_df

## Alias - Threat Fox
### Data Cleaning

#### Error Correction


alias_df = pd.read_csv(alias, low_memory = False )
alias_df

Threat_Fox_ALIAS_df = alias_df.dropna().reset_index(drop = True)

#### Data Normalization 

All the chars of tag are lower cases and all the punctualization factors are deleted

Threat_Fox_ALIAS_df = prepare_for_similarity_comparison(Threat_Fox_ALIAS_df, 'malware', 'malware')
Threat_Fox_ALIAS_df = prepare_for_similarity_comparison(Threat_Fox_ALIAS_df, 'alias', 'alias')

Threat_Fox_ALIAS_df

#### Data Deduplications 

Drop duplicates

Threat_Fox_ALIAS_df = Threat_Fox_ALIAS_df.drop_duplicates().reset_index(drop = True)
Threat_Fox_ALIAS_df

# SAVE into ..\IocSegnalations\PostDataCleaning\Threat_Fox\ folder


dataframes_dict = {}

# Add the DataFrames to the dictionary with keys
dataframes_dict['Entries']  = Threat_Fox_ENTRIES_df
dataframes_dict['Tags']     = Threat_Fox_TAGS_df
dataframes_dict['Malwares'] = Threat_Fox_MALWARES_df
dataframes_dict['Alias']    = Threat_Fox_ALIAS_df

save_directory = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Threat_Fox\\"
for key, df in dataframes_dict.items():
    file_name = f"{key}.csv"
    file_path = save_directory + file_name
    df.to_csv(file_path, index=False)
