==> 04.4. Data Integration .ipynb <==
## 4. Data Integration 

Integration between Threat Fox, Feodo Tracker and URL Haus datasets

## Divide the entries into multiple tables

- Tags
- Malware
- Entries
- Alias


import datetime
import os
from OTXv2 import OTXv2
import time
import json 
import pandas as pd 
import requests
import urllib3
import json
import sys


pd.options.display.max_colwidth = 500

# THREAT FOX

entries_TF_dir  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Threat_Fox\\Entries.csv"
malwares_TF_dir = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Threat_Fox\\Malwares.csv"
tags_TF_dir     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Threat_Fox\\Tags.csv"
alias_TF_dir    = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Threat_Fox\\Alias.csv"

### Tags - Threat Fox

tags_TF = pd.read_csv(tags_TF_dir, low_memory = False)
tags_TF

### ALIAS - Threat Fox

alias_TF = pd.read_csv(alias_TF_dir, low_memory = False)
alias_TF

### ENTRIES - Threat Fox

entries_TF = pd.read_csv(entries_TF_dir, low_memory = False)
entries_TF

### MALWARES - Threat Fox

malwares_TF = pd.read_csv(malwares_TF_dir, low_memory = False)
malwares_TF

# FEODO TRACKER


entries_FT_dir  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Feodo_Tracker\\Entries.csv"
malwares_FT_dir = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\Feodo_Tracker\\Malwares.csv"

### ENTRIES - FEODO TRACKER

entries_FT = pd.read_csv(entries_FT_dir, low_memory = False)
entries_FT

### MALWARES - FEODO TRACKER

malwares_FT = pd.read_csv(malwares_FT_dir, low_memory = False)
malwares_FT

# URL Haus 

entries_UH_dir  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\URL_Haus\\Entries.csv"
tags_UH_dir     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataCleaning\\URL_Haus\\Tags.csv"

### ENTRIES - URL Haus

entries_UH = pd.read_csv(entries_UH_dir, low_memory = False )
entries_UH 

### TAGS - URL Haus

tags_UH = pd.read_csv(tags_UH_dir, low_memory = False )
tags_UH 

## FOLLOWING THE INTEGRATION SCHEMA PERFORMED ON PAPER

Final dataset has the following fields:

- ENTRIES: final_df ( // ID_ENTRY , source // , 'ioc', 'threat_type', 'ioc_type', 'malware', 'reference', 'first_seen', 'last_seen', 'reporter',  )
       
- TAGS:  final_tags_df( // ID_ENTRY, source , tag_name // )

- ALIAS: final_MALWARES_df ( // malware // , alias )

- MALWARES: final_MALWARES_df ( // malware //)


final_ENTRIES_df = pd.DataFrame()

final_ENTRIES_df = pd.concat([entries_FT, entries_TF, entries_UH], ignore_index= True)

final_ENTRIES_df

## TAGS

final_TAGS_df = pd.DataFrame()

final_TAGS_df = pd.concat([tags_UH, tags_TF], ignore_index= True)

final_TAGS_df

### MALWARES

final_MALWARES_df = pd.DataFrame()

final_MALWARES_df = pd.concat([malwares_FT, malwares_TF], ignore_index= True)

final_MALWARES_df

### ALIAS

final_ALIAS_df = pd.DataFrame()

final_ALIAS_df = pd.concat([alias_TF], ignore_index= True)

final_ALIAS_df

# SAVE into ..\IocSegnalations\PostDataIntegration\ folder

dataframes_dict = {}

# Add the DataFrames to the dictionary with keys
dataframes_dict['Entries'] = final_ENTRIES_df
dataframes_dict['Tags'] = final_TAGS_df
dataframes_dict['Malwares'] = final_MALWARES_df
dataframes_dict['Alias'] = final_ALIAS_df


save_directory = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\IocSegnalations\\PostDataIntegration\\"
for key, df in dataframes_dict.items():
    file_name = f"{key}.csv"
    file_path = save_directory + file_name
    df.to_csv(file_path, index=False)
