==> 05.1. Data Profiling - Entries.ipynb <==
# 5.1 Data Profiling - Integrated Dataset - Entries

import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

file = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataIntegration\Entries.csv'

df = pd.read_csv(file, low_memory = False )
df

print("Duplicates: ",df.duplicated().any())
#total number of not null values
not_null_values = df.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = df.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = df.shape[0]*df.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN']
ds_property_nan = pd.read_csv(file, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

## Completeness

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print(completeness)

# Single column analysis

# 1 : ioc

## CARDINALITIES

#len command counts the number of rows
n_rows = len(df['ioc'])
print("# rows: ", n_rows)

#number of non—null observation in "ioc" column
count_values = df['ioc'].count()
print("# non-null: ", n_rows)

#value counts is the same of a group by (attribute "ibu")
df['ioc'].value_counts()

## Cardinality

#nunique is the number of distict values for an attribute (attribute "ibu")
distinct_values = (df['ioc'].nunique())
print("Cardinality: ", distinct_values)

## Uniqueness

#Uniqueness evaluation (attribute "ioc")
uniqueness = distinct_values / n_rows
print("Uniqueness: ", uniqueness)

## Distinctness

#Distinctness evaluation (attribute "ioc")
distinctness = distinct_values / count_values
print("Distinctness: ", distinctness)

# 2 : threat_type


## CARDINALITIES


col = 'threat_type'

n_rows = len(df[col])
print("# rows: ", n_rows)

count_values = df[col].count()
print("# non-null: ", n_rows)

df[col].value_counts()

## Cardinality

distinct_values = (df[col].nunique())
print("Cardinality: ", distinct_values)

## Uniqueness

uniqueness = distinct_values / n_rows
print("Uniqueness: ", uniqueness)

## Distinctness

distinctness = distinct_values / count_values
print("Distinctness: ", distinctness)

# 2 : ioc_type


## CARDINALITIES


col = 'ioc_type'

n_rows = len(df[col])
print("# rows: ", n_rows)

count_values = df[col].count()
print("# non-null: ", n_rows)

df[col].value_counts()

## Cardinality

distinct_values = (df[col].nunique())
print("Cardinality: ", distinct_values)

## Uniqueness

uniqueness = distinct_values / n_rows
print("Uniqueness: ", uniqueness)

## Distinctness

distinctness = distinct_values / count_values
print("Distinctness: ", distinctness)

# 3 : malware


## CARDINALITIES


col = 'malware'

n_rows = len(df[col])
print("# rows: ", n_rows)

count_values = df[col].count()
print("# non-null: ", n_rows)

df[col].value_counts()

## Cardinality

distinct_values = (df[col].nunique())
print("Cardinality: ", distinct_values)

## Uniqueness

uniqueness = distinct_values / n_rows
print("Uniqueness: ", uniqueness)

## Distinctness

distinctness = distinct_values / count_values
print("Distinctness: ", distinctness)

# 4 : reporter

## CARDINALITIES


col = 'reporter'

n_rows = len(df[col])
print("# rows: ", n_rows)

count_values = df[col].count()
print("# non-null: ", n_rows)

df[col].value_counts()

## Cardinality

distinct_values = (df[col].nunique())
print("Cardinality: ", distinct_values)

## Uniqueness

uniqueness = distinct_values / n_rows
print("Uniqueness: ", uniqueness)

## Distinctness

distinctness = distinct_values / count_values
print("Distinctness: ", distinctness)

# Profiling Auto-report

pip install pandas-profiling

pip install --upgrade numba

pip install visions==0.7.4

pip install --upgrade pydantic pydantic-settings

pip install ydata-profiling

from ydata_profiling import ProfileReport

df['first_seen'] = pd.to_datetime(df['first_seen'], format='mixed', utc = True)
df['last_seen'] = pd.to_datetime(df['last_seen'], format='mixed', utc = True)

df = df.dropna(subset=['first_seen', 'last_seen'])

profile = ProfileReport(df, title="Integrated Dataset - Entries - Pandas Profiling Report")

profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 5\\IntegratedDataset\\Entries.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 5\\IntegratedDataset\\Entries.json")

## Value Distribution

Histograms representing the source and the timeline of the data

#### Distribution of malwares

#import libraries and set up
import seaborn as sns
import matplotlib
matplotlib.use('tkagg')
import matplotlib.pyplot as plt


sns.set(color_codes=True)
sns.set_palette(sns.color_palette("muted"))

df['first_seen'] = pd.to_datetime(df['first_seen'], utc=True, format='mixed')

threat_fox_df = df[ df['source'] == 'Threat Fox' ]
url_haus_df = df[ df['source'] == 'URL Haus' ]
feodo_tracker_df = df[ df['source'] == 'Feodo Tracker' ]

def distibution_plot(df, field):
    
    unique_malwares = df[field].unique()
    
    new = pd.DataFrame()
    
    s = 0
    
    nunique = len(unique_malwares)
    
    for i in range(nunique):
        
        malware = unique_malwares[i]
        
        s = s + len(df[df[field] == malware ])
    
    avg = ( s / nunique )
    
    for i in range(nunique):
        
        malware = unique_malwares[i]
        
        tmp = df[df[field] == malware ]
    
        if(len(tmp) > avg):
            
            new = pd.concat([new, tmp])       
            
    return new       

distibution_plot(df, 'malware')['malware'].hist()

#### Distribution of entries over time

# Create a time series plot with different colors for each category
plt.figure(figsize=(10, 6))

# Plot Category A with a blue color
plt.plot(threat_fox_df['first_seen'], label='Threat Fox', marker='o', linestyle='-', color='blue')

# Plot Category B with a green color
plt.plot(url_haus_df['first_seen'], label='URL Haus', marker='x', linestyle='-', color='green')

# Plot Category C with a red color
plt.plot(feodo_tracker_df['first_seen'], label='Feodo Tracker', marker='s', linestyle='-', color='red')

# Customize the plot
plt.title('Time Series Data by Category')
plt.xlabel('Timestamp')
plt.grid(True)

# Add a legend to distinguish between categories
plt.legend()

# Show the plot
plt.tight_layout()
plt.show()
