==> 06.1. Data Cleaning - Entries.ipynb <==
# 6. Data Cleaning - Entries

Thanks to the previous step, Data Profiling - Entries



import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np
import math

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

file = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataIntegration\Entries.csv'

df = pd.read_csv(file, low_memory = False )
df

# Data Cleaning
### Handle missing values

We notice that there is a HIGH number of missing values columns (<=50%) so i have decided to drop those columns which are:

- status

column_name = 'status'
completeness_percentage = (df[column_name].count() / len(df)) * 100

print(f"Completeness of {column_name}: {completeness_percentage:.2f}%")

df = df.drop(columns = column_name)
df

In 'ioc_type' related to URL Haus as source are all 'NaN'. So, i set them to url because the URL Haus provide only url segnalitions

column_name = 'ioc_type'
completeness_percentage = (df[column_name].count() / len(df)) * 100

print(f"Completeness of {column_name}: {completeness_percentage:.2f}%")

tmp = df[df['source'] == 'URL Haus']['ioc_type']
tmp_missings = tmp.isna().count()
print(f"Percentage of missings of {column_name}: {(tmp_missings / len(tmp) ) * 100:.2f}%")

df.loc[df['source'] == 'URL Haus','ioc_type'] = 'url'
df

### Standardization : Field 'threat_type'

 'payload delivery' from 'payload_delivery' in the dataframe

df['threat_type'].unique()

df['threat_type'] = df['threat_type'].replace('payload', 'payload delivery').replace('payload_delivery', 'payload delivery')
df

### Normalization

Date attribute as 'first_seen' and 'last_seen' have multiple format so i have decided to standardize it to the minimal one that all of them have in common:  
- year - month - day  

standardized_df = df

from datetime import datetime, timedelta

df['last_seen'] = pd.to_datetime(df['last_seen'], utc=True, format='mixed').dt.strftime("%Y-%m-%d")
df['first_seen'] = pd.to_datetime(df['first_seen'], utc=True, format='mixed').dt.strftime("%Y-%m-%d")
df


#### Missing value
Drop if one of the keys 'ID_ENTRY', 'source' are NaN

df = df.dropna(subset=['ID_ENTRY', 'source'])
df = df.reset_index(drop = True)
df

# What about malware missings?

They will be managed in Data Enrichment using Tags, Malwares and Alias tables

### Save dataset into DataWrangling folder

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\IntegratedDataset\PostDataCleaning\Entries.csv')
