==> 06.2 Data Cleaning - Tags.ipynb <==
# 06.2 Data Cleaning - Tags

Thanks to the previous step, Data Profiling - Tags



import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np
import math

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

file = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataIntegration\Tags.csv'

df = pd.read_csv(file, low_memory = False )
df

# Data Cleaning

### Error Correction - Missing value
Drop a row if one of the sub columns 'tag' , 'source' or 'ID_ENTRY' is NaN

df.dropna(subset=['ID_ENTRY', 'tag', 'source'], inplace=True)
df = df.reset_index(drop = True)
df

### Duplicate Detection - Drop Duplicates
The duplicate rows are dropped

df = df.drop_duplicates()
df = df.reset_index(drop = True)
df

### Save dataset into DataWrangling folder

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\IntegratedDataset\PostDataCleaning\Tags.csv')
