==> 06.3 Data Cleaning - Alias.ipynb <==
# 06.3 Data Cleaning - Alias

Thanks to the previous step, Data Profiling - Alias



import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np
import math

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

file = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\PostDataIntegration\Alias.csv'

df = pd.read_csv(file, low_memory = False )
df

# Data Cleaning

### Missing value 
- Drop a row if one of the sub columns 'malware' , 'alias' is NaN
- Drop a row if it's duplicate

df.dropna(subset=['malware', 'alias'], inplace=True)
df = df.reset_index(drop = True)
df

df = df.drop_duplicates()
df = df.reset_index(drop = True)

df

### Save dataset into DataWrangling folder

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\IocSegnalations\IntegratedDataset\PostDataCleaning\Alias.csv')
