==> 11.1. Data Enrichment - OTX Alien Vault - ip addresses.ipynb <==
# 11. Data Enrichment

from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

file = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\DataCollection\OTXAlienVault\file_relativeTo_ip.csv'

df = pd.read_csv(file, low_memory = False)
df

df = df.drop('Unnamed: 0', axis=1)
df

# Data Cleaning

## 1. Datetype conversion 

df = df.drop(['date'], axis=1)
df['date'] = pd.to_datetime(df['datetime_int'], unit='s')
df = df.drop('datetime_int', axis=1)

df

# Data Deduplication

def find_duplicates(input_list):
    seen = {}
    duplicates = []
    for item in input_list:
        if item in seen:
            duplicates.append(item)
        else:
            seen[item] = 1
    return duplicates

## Field analyzed: 'hash'

df['hash'].unique()

files = df[df['hash'].notna()]
files 

dup_files = find_duplicates(files.hash)
len(dup_files)

## Summary

d = dup_files[0]

df[df['hash'] == d]

# By a first look at the duplicates in the DB, it seems that there are 
# multiple ip addresses related to the same hash with all the other fields in common

# This function will return true if this assumption will be correct

def analyze_duplicates(df, duplicates):
    
    hp = True
    x = []
    
    for i in range(len(duplicates)):
        
        el = duplicates[i]
        
        dataframe = df[df['hash'] == el]
        dataframe.reset_index(drop=True, inplace=True)
                        
        tmp_avast      = str(dataframe.at[0, 'detections.avast'])
        tmp_clamav     = str(dataframe.at[0, 'detections.clamav'])
        tmp_msdefender = str(dataframe.at[0, 'detections.msdefender'])
        tmp_date       = str(dataframe.at[0, 'date'])
        tmp_ip_address = str(dataframe.at[0, 'ip address'])
                
        for j in range(len(dataframe) - 1):
            
            if(str(dataframe.at[j + 1 , 'date']) != tmp_date ):
                print(f"{dataframe.at[j + 1 , 'date']} and {tmp_date} are different")
                
                hp = False
                
            if(str(dataframe.at[j + 1 , 'detections.msdefender']) != tmp_msdefender ):
                print(f"{dataframe.at[j + 1 , 'detections.msdefender']} and {tmp_msdefender} are different")

                hp = False
                
            if(str(dataframe.at[j + 1 , 'detections.avast']) != tmp_avast ):
                print(f"{dataframe.at[j + 1 , 'detections.avast']} and {tmp_avast} are different")

                hp = False
                
            if(str(dataframe.at[j + 1 , 'detections.clamav']) != tmp_clamav ):
                print(f"{dataframe.at[j + 1 , 'detections.clamav']} and {tmp_clamav} are different")

                hp = False
            
    return hp
    
    

print(f" The previous hypothesys was correct?  {analyze_duplicates(df, dup_files)}")

So, talking about the ER representations. Both the fields 'ip address' and 'hash' are Keys.

# Data Enrichment

(Probably this could be done at the end of the pipeline?) In this field i will take into account the 3 fields related to detections and i'll predict a new field 

# Save Dataset

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\DataCollection_OTXAlienVault\hash_ips.csv')
