==> 12. Data Integration - Hash .ipynb <==
# 12.0. Data Integration 

- Use '..\Pipeline\BeforeCollection\files.csv' : so hashes from URL Haus, Feodo Tracker and Threat Fox.
- Use '..\Pipeline\DataCollection_OTXAlienVault\hash_domain.csv' : (hash, domains) collected from OTXAlienVault
- Use '..\Pipeline\DataCollection_OTXAlienVault\hash_ips.csv' : (hash, ip address) collected from OTXAlienVault

to perform a look up on the Malware Bazaar in order to add an analysis 

from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

file = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\BeforeCollection\files.csv'

ipFiles = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\DataCollection_OTXAlienVault\hash_ips.csv'

urlFiles = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\DataCollection_OTXAlienVault\hash_domain.csv'

df_files = pd.read_csv(file, low_memory = False)
df_files

df_ip_files = pd.read_csv(ipFiles, low_memory = False)
df_ip_files

df_url_files = pd.read_csv(urlFiles, low_memory = False)
df_url_files

# Join Hashes in a list

lista_hash =  list(set(list(set(df_url_files['hash'])) + list(set(df_ip_files['hash'])) + list(set(df_files['ioc']))))
lista_hash

len(lista_hash)

df = pd.DataFrame()
df['hash'] = lista_hash
df

df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\DataIntegration_OTXAlienVault\hash.csv')
