==> 15.0. Data Profiling - OTX Alien Vault - Hash Analysis Dataset.ipynb <==
# 15.0. Data Profiling - OTX Alien Vault

from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

pip install pandas-profiling

pip install --upgrade numba

pip install visions==0.7.4

pip install --upgrade pydantic pydantic-settings

pip install ydata-profiling

from ydata_profiling import ProfileReport

sys.path.append("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\")

### Data Profiling - ENTRIES

path_Entries = r"C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\DataAnalysis\otx_alien_vault\Entries.csv"

entries = pd.read_csv(path_Entries, low_memory = False)
entries

# Define new column names
columns = ['analysis.info.results.sha1', 'analysis.info.results.sha256', 'analysis.info.results.md5', 'analysis.datetime_int', 'page_type']
entries_df = entries[columns]

print("Duplicates: ",entries.duplicated().any())
#total number of not null values
not_null_values = entries.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = entries.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = entries.shape[0]*entries.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Entries, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

completeness = 1 - entries['analysis.info.results.sha1'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'analysis.info.results.sha1': {completeness * 100:.2f}%")

completeness = 1 - entries['analysis.info.results.sha256'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'analysis.info.results.sha256': {completeness * 100:.2f}%")

completeness = 1 - entries['analysis.info.results.md5'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'analysis.info.results.md5': {completeness * 100:.2f}%")

completeness = 1 - entries['analysis.datetime_int'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'analysis.datetime_int': {completeness * 100:.2f}%")

completeness = 1 - entries['page_type'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'page_type': {completeness * 100:.2f}%")

profile = ProfileReport(entries, title="OTX Alien Vault - IP Addresses - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTX_Alien_Vault-IP_Addresses.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTX_Alien_Vault-IP_Addresses.json")

### Data Profiling - ANOMALIES

path_Anomalies = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\DataAnalysis\\otx_alien_vault\\Anomalies.csv"

anomalies = pd.read_csv(path_Anomalies, low_memory = False)
anomalies.head(3)

print("Duplicates: ",anomalies.duplicated().any())
#total number of not null values
not_null_values = anomalies.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = anomalies.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = anomalies.shape[0]*anomalies.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Anomalies, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

profile = ProfileReport(anomalies, title="OTX Alien Vault - Hash Analysis - Anomalies Report")
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_AnomaliesReport.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_AnomaliesReport.json")

### Data Profiling - ATTACKS

path_Attacks = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\DataAnalysis\\otx_alien_vault\\Attacks.csv"

attacks = pd.read_csv(path_Attacks, low_memory = False)
attacks.head(3)

print("Duplicates: ",attacks.duplicated().any())
#total number of not null values
not_null_values = attacks.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = attacks.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = attacks.shape[0]*attacks.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Attacks, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

profile = ProfileReport(attacks, title="OTX Alien Vault - Hash Analysis - Attacks Report")
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_AttacksReport.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_AttacksReport.json")

### Data Profiling - BEHAVIOURS

path_Behaviours = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\DataAnalysis\\otx_alien_vault\\Behaviours.csv"

behaviours = pd.read_csv(path_Behaviours, low_memory = False)
behaviours.head(3)

print("Duplicates: ",behaviours.duplicated().any())
#total number of not null values
not_null_values = behaviours.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = behaviours.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = behaviours.shape[0]*behaviours.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Behaviours, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

completeness = 1 - behaviours['sha1'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'sha1': {completeness * 100:.2f}%")

completeness = 1 - behaviours['sha256'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'sha256': {completeness * 100:.2f}%")

completeness = 1 - behaviours['md5'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'md5': {completeness * 100:.2f}%")

completeness = 1 - behaviours['name'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'name': {completeness * 100:.2f}%")

completeness = 1 - behaviours['description'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'description': {completeness * 100:.2f}%")

completeness = 1 - behaviours['severity'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'severity': {completeness * 100:.2f}%")

completeness = 1 - behaviours['markcount'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'markcount': {completeness * 100:.2f}%")

profile = ProfileReport(behaviours, title="OTX Alien Vault - Hash Analysis - Behaviours Report")
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_BehavioursReport.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_BehavioursReport.json")

### Data Profiling - THREAT_VENDOR_ANALYSIS

path_TV_Analysis = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\DataAnalysis\\otx_alien_vault\\TV_Analysis.csv"

tv_analysis = pd.read_csv(path_TV_Analysis, low_memory = False)
tv_analysis.head(3)

print("Duplicates: ",tv_analysis.duplicated().any())
#total number of not null values
not_null_values = tv_analysis.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = tv_analysis.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = tv_analysis.shape[0]*tv_analysis.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_TV_Analysis, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

completeness = 1 - tv_analysis['sha1'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'sha1': {completeness * 100:.2f}%")

completeness = 1 - tv_analysis['sha256'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'sha256': {completeness * 100:.2f}%")

completeness = 1 - tv_analysis['md5'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'md5': {completeness * 100:.2f}%")

completeness = 1 - tv_analysis['malware'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'malware': {completeness * 100:.2f}%")

completeness = 1 - tv_analysis['verdict'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'verdict': {completeness * 100:.2f}%")

completeness = 1 - tv_analysis['threat_vendor_name'].isna().mean()

# Print the completeness percentage
print(f"Completeness for 'threat_vendor_name': {completeness * 100:.2f}%")

profile = ProfileReport(tv_analysis, title="OTX Alien Vault - Hash Analysis - Threat Vendor Analysis Report")
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_ThreatVendorAnalysisReport.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 15\\OTX_Alien_Vault\\OTXAlienVault_ThreatVendorAnalysisReport.json")
