==> 17.0. Data Integration - File Analysis.ipynb <==
# 17.0. Data Integration - File Analysis

import requests
import json

# Replace 'YOUR_API_KEY' with your actual Malware Bazaar API key
api_key = ''

# URL of the Malware Bazaar API
api_url = 'https://mb-api.abuse.ch/api/v1/'

# SHA-256 hash of the file
file_sha256 = 'eb0e81598d8526d88cac4695a3e9360cc8fbb331'

# Make a query to the Malware Bazaar API
response = requests.post(api_url, data={'query': 'get_info', 'hash': file_sha256,})

# Check if the request was successful (status code 200)
if response.status_code == 200:
    # Parse the JSON response
    result = response.json()
    formatted_json = json.dumps(result, indent=2)

    print(formatted_json)
else:
    # Print an error message if the request was not successful
    print(f"Error: {response.status_code} - {response.text}")


from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

path_OTX_Entries     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\OTX_Alien_Vault\\Entries.csv"
path_OTX_Anomalies   = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\OTX_Alien_Vault\\Anomalies.csv"
path_OTX_Attacks     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\OTX_Alien_Vault\\Attacks.csv"
path_OTX_TV_Analysis = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\OTX_Alien_Vault\\TV_Analysis.csv"
path_OTX_Behaviours  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\OTX_Alien_Vault\\Behaviours.csv"

path_MB_Entries     = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Malware_Bazaar\\Entries.csv"
path_MB_Tags        = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Malware_Bazaar\\Tags.csv"
path_MB_YaraRules   = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Malware_Bazaar\\YaraRules.csv"
path_MB_TV_Analysis = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Malware_Bazaar\\TV_Analysis.csv"
path_MB_Behaviours  = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Malware_Bazaar\\Behaviours.csv"

## Entries

entries_MB = pd.read_csv(path_MB_Entries, low_memory = False)
left_key = ['sha1', 'sha256', 'md5']
entries_MB.head(3)

entries_OTX = pd.read_csv(path_OTX_Entries, low_memory = False)
right_key = ['sha1', 'sha256', 'md5']
entries_OTX.head(3)

entries = pd.merge(entries_MB, entries_OTX, left_on=left_key, right_on=right_key, how='outer')
entries

#### Column Merging

- merge column 'file_type_x' and 'file_type_y'


# Verifica se, per tutte le righe, la colonna 'a' è uguale alla colonna 'b' oppure almeno una delle due è NaN
condition_for_all_rows = (entries.apply(lambda row: pd.isna(row['file_type_x']) or pd.isna(row['file_type_y']) or row['file_type_x'] == row['file_type_y'], axis=1)).all()

# Visualizza il risultato
print(f"La condizione è vera per tutte le righe: {condition_for_all_rows}")


entries['file_type'] = entries['file_type_x'].combine_first(entries['file_type_y']).combine_first(entries['file_type_x'])

entries = entries.drop(columns=['file_type_x', 'file_type_y'])
entries

ordered_entries_columns = ['sha1', 'sha256', 'md5', 'first_seen', 'last_seen', 'analysis_date', 'file_type', 'delivery_method' ]
entries = entries[ordered_entries_columns]
entries.head(3)

## Threat Vendors Analysis

tv_analysis_MB = pd.read_csv(path_MB_TV_Analysis, low_memory = False)
tv_analysis_MB.head(3)

tv_analysis_OTX = pd.read_csv(path_OTX_TV_Analysis, low_memory = False)
tv_analysis_OTX.head(3)

tv_analysis = pd.concat([tv_analysis_MB, tv_analysis_OTX])
tv_analysis.reset_index(drop=True, inplace=True)
tv_analysis

## Behaviours

behaviours_MB = pd.read_csv(path_MB_Behaviours, low_memory = False)
behaviours_MB.rename(columns= { 'action_maliciousness_level' : 'maliciousness_level' }, inplace=True)

behaviours_MB.head(3)

behaviours_OTX = pd.read_csv(path_OTX_Behaviours, low_memory = False)
behaviours_OTX.rename(columns= { 'action_level' : 'maliciousness_level' }, inplace=True)

behaviours_OTX.head(3)

behaviours = pd.concat([behaviours_OTX, behaviours_MB])
behaviours

## Yara Rules

yara_rules = pd.read_csv(path_MB_YaraRules, low_memory = False)
yara_rules.head(3)

## Anomalies

anomalies = pd.read_csv(path_OTX_Anomalies, low_memory = False)
anomalies

## Attacks

attacks = pd.read_csv(path_OTX_Attacks, low_memory = False)
attacks

## Tags

tags = pd.read_csv(path_MB_Tags, low_memory = False)
tags.head(3)

# Save dataset

dataframes_dict = {}

# Add the DataFrames to the dictionary with keys
dataframes_dict['Entries'] = entries
dataframes_dict['TV_Analysis'] = tv_analysis
dataframes_dict['Yara_Rules'] = yara_rules
dataframes_dict['Tags'] = tags
dataframes_dict['Behaviours'] = behaviours
dataframes_dict['Anomalies'] = anomalies
dataframes_dict['Attacks'] = attacks

save_directory = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\"
for key, df in dataframes_dict.items():
    file_name = f"{key}.csv"
    file_path = save_directory + file_name
    df.to_csv(file_path, index=False)
    print(f'{key} added')
