==> 18.0. Data Profiling - Integrated Dataset - File Analysis.ipynb <==
# 18.0. Data Profiling - Integrated Dataset 

from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

from ydata_profiling import ProfileReport

### Data Profiling - ENTRIES

path_Entries = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\Entries.csv"

entries = pd.read_csv(path_Entries, low_memory = False)
entries

print("Duplicates: ",entries.duplicated().any())
#total number of not null values
not_null_values = entries.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = entries.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = entries.shape[0]*entries.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Entries, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

def analyse_columns(df):
    
    columns = df.columns
    for i in range(len(columns)):
        
        compute_completeness(df, columns[i])

def compute_completeness(df, col):
    completeness = 1 - df[col].isna().mean()

    # Print the completeness percentage
    print(f"Completeness for '{(col)}': {completeness * 100:.2f}%")

analyse_columns(entries)

profile = ProfileReport(entries, title="Entries - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Entries_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Entries_Report.json")

### Data Profiling - ANOMALIES

path_Anomalies = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\Anomalies.csv"

anomalies = pd.read_csv(path_Anomalies, low_memory = False)
anomalies.head(3)

print("Duplicates: ",anomalies.duplicated().any())
#total number of not null values
not_null_values = anomalies.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = anomalies.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = anomalies.shape[0]*anomalies.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Anomalies, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

analyse_columns(anomalies)

profile = ProfileReport(anomalies, title="Anomalies - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Anomalies_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Anomalies_Report.json")

### Data Profiling - ATTACKS

path_Attacks = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\Attacks.csv"

attacks = pd.read_csv(path_Attacks, low_memory = False)
attacks.head(3)

print("Duplicates: ",attacks.duplicated().any())
#total number of not null values
not_null_values = attacks.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = attacks.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = attacks.shape[0]*attacks.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Attacks, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

analyse_columns(attacks)

profile = ProfileReport(attacks, title="Attacks - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Attacks_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Attacks_Report.json")

### Data Profiling - BEHAVIOURS

path_Behaviours = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\Behaviours.csv"

behaviours = pd.read_csv(path_Behaviours, low_memory = False)
behaviours.head(3)

print("Duplicates: ",behaviours.duplicated().any())
#total number of not null values
not_null_values = behaviours.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = behaviours.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = behaviours.shape[0]*behaviours.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Behaviours, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

analyse_columns(behaviours)

profile = ProfileReport(behaviours, title="Behaviours - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Behaviours_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Behaviours_Report.json")

### Data Profiling - THREAT_VENDOR_ANALYSIS

path_TV_Analysis = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\TV_Analysis.csv"

tv_analysis = pd.read_csv(path_TV_Analysis, low_memory = False)
tv_analysis.head(3)

print("Duplicates: ",tv_analysis.duplicated().any())
#total number of not null values
not_null_values = tv_analysis.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = tv_analysis.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = tv_analysis.shape[0]*tv_analysis.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_TV_Analysis, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

analyse_columns(tv_analysis)

profile = ProfileReport(tv_analysis, title="Threat Vendor Analysis - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\TV_Analysis_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\TV_Analysis_Report.json")

### Data Profiling - TAGS

path_Tags = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\Tags.csv"

tags = pd.read_csv(path_Tags, low_memory = False)
tags.head(3)

print("Duplicates: ",tags.duplicated().any())
#total number of not null values
not_null_values = tags.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = tags.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = tags.shape[0]*tags.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_Tags, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

analyse_columns(tags)

profile = ProfileReport(tags, title="Tags - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Tags_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Tags_Report.json")

### Data Profiling - Yara Rules

path_yaraRules = r"C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataIntegration\\Yara_Rules.csv"

yaraRules = pd.read_csv(path_yaraRules, low_memory = False)
yaraRules.head(3)

print("Duplicates: ",yaraRules.duplicated().any())
#total number of not null values
not_null_values = yaraRules.count().sum()
print("# not null values: ",not_null_values)
#total number of null values
null_values = yaraRules.isnull().sum().sum()
print("# null values: ",null_values)
#total number of cells
n_values = yaraRules.shape[0]*yaraRules.shape[1]
print("# values: ",n_values)
#we added to the set of missing values also 'na' and '--'
missing_values = ['na','--', 'NaN', 'None', '']
ds_property_nan = pd.read_csv(path_yaraRules, na_values = missing_values)
print("# missings: ",ds_property_nan.isnull().sum().sum())

#COMPLETENESS evaluation
completeness = not_null_values / n_values
completeness = '{0:.1f}%'.format(completeness*100)
print("Completeness: ",completeness)

analyse_columns(yaraRules)

profile = ProfileReport(yaraRules, title="Yara Rules - Pandas Profiling Report", minimal=True)
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Yara_Rules_Report.html")

# As a JSON string
json_data = profile.to_json()

# As a file
profile.to_file("C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Report\\Profiling step 18\\Integrated_Dataset\\Yara_Rules_Report.json")
