==> 20.0. Data Enrichment - Integrated Dataset - File Analysis.ipynb <==
# 20.0. Data Enrichment - Integrated Dataset - File Analysis

pip install OTXv2

entries_dir      = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Entries.csv"
attacks_dir      = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Attacks.csv"
anomalies_dir    = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Anomalies.csv"
tv_analysis_dir  = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\TV_Analysis.csv"
behaviours_dir   = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Behaviours.csv"
yaraRules_dir    = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Yara_Rules.csv"
tags_dir         = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Tags.csv"

from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np
import pandas as pd
from OTXv2 import OTXv2

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

api_key = '006ee711b0794e449a6d074aacdda04a285032c7e787893b095cc8a2308f99ee'
otx = OTXv2('006ee711b0794e449a6d074aacdda04a285032c7e787893b095cc8a2308f99ee')

## Entries Enrichment 
Add the # pulses in which this file is reported

entries_df = pd.read_csv(entries_dir, low_memory = False)
entries_df

#### Example

i = 9
h = entries_df.at[i, 'sha256']
sha256 = h
sha1 = entries_df.at[i, 'sha1']


print(sha256)
print(sha1)

otx_url = 'https://otx.alienvault.com/api/v1/indicators/file/'+ h

current_pulse = requests.get(otx_url,headers={"X-OTX-API-KEY":api_key})

current_pulse = json.loads(current_pulse.content.decode("utf-8", "ignore"))

tmp = pd.DataFrame.from_dict(pd.json_normalize(current_pulse)).reset_index(drop = True)

tmp



 

#### Enrichment process
Add the field '#OTX_pulses

def add_otx_pulses_number(df):
    
    max_retries = 5
    
    retry_delay= 10
    
    for i in range(len(df)):
        
        print(i)
        
        sha256 = df.at[i, 'sha256']
        sha1 = df.at[i, 'sha1']
        
        first_seen = df.at[i, 'first_seen']
        last_seen = df.at[i, 'last_seen']
        
        otx_url = 'https://otx.alienvault.com/api/v1/indicators/file/'+ str(sha256)
        
        try:
            for attempt in range(1, max_retries + 1):

                current_pulse = requests.get(otx_url,headers={"X-OTX-API-KEY":api_key})

                current_pulse = json.loads(current_pulse.content.decode("utf-8", "ignore"))

                tmp = pd.DataFrame.from_dict(pd.json_normalize(current_pulse)).reset_index(drop = True)
                
                if 'pulse_info.count' in tmp.columns:
                
                    if( str(sha256) == tmp.at[0,'indicator']):

                        if 'pulse_info.count' in tmp.columns:

                            c = tmp.at[0, 'pulse_info.count']

                            df.loc[i,'#OTX_pulses'] = c

                        if (pd.isna(first_seen) or pd.isna(last_seen)) and 'pulse_info.pulses' in tmp.columns:

                            tmp = pd.json_normalize(pd.DataFrame.from_dict(pd.json_normalize(tmp['pulse_info.pulses']).T).reset_index(drop = True)[0])

                            if(len(tmp) > 0):

                                tmp['created'] = pd.to_datetime(tmp['created'],  format='mixed').dt.strftime('%Y-%m-%d')

                                oldest_date = tmp['created'].min()
                                df.loc[i,'first_seen'] = oldest_date

                                newest_date = tmp['created'].max()
                                df.loc[i,'last_seen'] = newest_date
                        
                
        except json.JSONDecodeError as e:
            print(f"JSON Decode Error on attempt {attempt}/{max_retries}: {e}")
            if attempt < max_retries:
                print(f"Retrying in {retry_delay} seconds...")
                time.sleep(retry_delay)
                
            else:
                print("Max retries reached. Unable to handle JSON Decode Error.")
                df.loc[i,'#OTX_pulses'] = 0
                
        except (requests.exceptions.ReadTimeout, requests.exceptions.RequestException) as e:
            print(f"Request attempt {attempt} timed out. Retrying...")
            if attempt < max_retries:
                print(f"Retrying in {retry_delay} seconds...")
                time.sleep(retry_delay)

            else:
                print("Max retries reached. Unable to handle JSON Decode Error.")
                df.loc[i,'#OTX_pulses'] = 0
                
        
            
    return df


            

missing_first_seen = entries_df['first_seen'].isnull().sum()
missing_last_seen = entries_df['last_seen'].isnull().sum()
#missing_counting  = entries_df['#OTX_pulses'].isnull().sum()

# Calculate completeness
completeness_first_seen = 1 - (missing_first_seen / len(entries_df))
completeness_last_seen = 1 - (missing_last_seen / len(entries_df))
#completeness_counting = 1 - (missing_counting / len(entries_df))

# Display the results
print("Completeness for 'first_seen': {:.2%}".format(completeness_first_seen))
print("Completeness for 'last_seen': {:.2%}".format(completeness_last_seen))
#print("Completeness for '#OTX_pulses': {:.2%}".format(completeness_counting))

entries_df_pulses_n_added = add_otx_pulses_number(entries_df)
entries_df_pulses_n_added

missing_first_seen = entries_df_pulses_n_added['first_seen'].isnull().sum()
missing_last_seen = entries_df_pulses_n_added['last_seen'].isnull().sum()
missing_counting  = entries_df_pulses_n_added['#OTX_pulses'].isnull().sum()

# Calculate completeness
completeness_first_seen = 1 - (missing_first_seen / len(entries_df_pulses_n_added))
completeness_last_seen = 1 - (missing_last_seen / len(entries_df_pulses_n_added))
completeness_counting = 1 - (missing_counting / len(entries_df_pulses_n_added))

# Display the results
print("Completeness for 'first_seen': {:.2%}".format(completeness_first_seen))
print("Completeness for 'last_seen': {:.2%}".format(completeness_last_seen))
print("Completeness for '#OTX_pulses': {:.2%}".format(completeness_counting))

entries_df_pulses_n_added.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\FileAnalysis\PostDataEnrichment\Step1\Entries.csv')
