==> 20.0. Data Enrichment - Integrated Dataset - OTX Pulses.ipynb <==
# 20.0. Data Enrichment - Integrated Dataset - OTX Pulses

pip install OTXv2

entries_dir      = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Entries.csv"
attacks_dir      = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Attacks.csv"
anomalies_dir    = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Anomalies.csv"
tv_analysis_dir  = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\TV_Analysis.csv"
behaviours_dir   = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Behaviours.csv"
yaraRules_dir    = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Yara_Rules.csv"
tags_dir         = "C:\\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\FileAnalysis\\PostDataCleaning\\Integrated_Dataset\\Tags.csv"

from datetime import datetime
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np
import pandas as pd
from OTXv2 import OTXv2

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

api_key = '006ee711b0794e449a6d074aacdda04a285032c7e787893b095cc8a2308f99ee'
otx = OTXv2('006ee711b0794e449a6d074aacdda04a285032c7e787893b095cc8a2308f99ee')

## Entries Enrichment 
Add the # pulses in which this file is reported

entries_df = pd.read_csv(entries_dir, low_memory = False)
entries_df

#### Example

i = 9
h = '04c53261b1220a894a02f5ffb39cdfd73f93481c0b5c8106d21c91b20205c62d'
sha1 = '411535c9548f86034a6a9a603de521a86a3b5466'

tmp1 = pd.DataFrame()

otx_url = 'https://otx.alienvault.com/api/v1/indicators/file/'+ sha1

current_pulse = requests.get(otx_url,headers={"X-OTX-API-KEY":api_key})

current_pulse = json.loads(current_pulse.content.decode("utf-8", "ignore"))

tmp = pd.DataFrame.from_dict(pd.json_normalize(current_pulse)).reset_index(drop = True)

tmp

i = 9
h = '04c53261b1220a894a02f5ffb39cdfd73f93481c0b5c8106d21c91b20205c62d'
sha1 = '411535c9548f86034a6a9a603de521a86a3b5466'
md5 = '5b5a0225a2b7b9a054417662c733168a'

tmp1 = pd.DataFrame()

otx_url = 'https://otx.alienvault.com/api/v1/indicators/file/'+ sha1

current_pulse = requests.get(otx_url,headers={"X-OTX-API-KEY":api_key})

current_pulse = json.loads(current_pulse.content.decode("utf-8", "ignore"))

tmp = pd.DataFrame.from_dict(pd.json_normalize(current_pulse)).reset_index(drop = True)

if ('pulse_info.count' in tmp.columns) & ('pulse_info.pulses' in tmp.columns):

    if tmp.at[0,'pulse_info.count'] > 0:
        
        tmp1 = pd.json_normalize(pd.DataFrame.from_dict(pd.json_normalize(tmp['pulse_info.pulses']).T).reset_index(drop = True)[0])
        
tmp1     


 

usefull_columns = ['id', 'name', 'description', 'modified', 'TLP', 'upvotes_count', 'downvotes_count', 'author.username']

#### Data collection PULSES

def create_otx_pulses_tdf(df):
    
    max_retries = 5
    retry_delay= 10
    dataframe = pd.DataFrame()
    
    for i in range(len(df)):
        
        print(i)
        
        sha256 = df.at[i, 'sha256']
        sha1 = df.at[i, 'sha1']
        md5 = df.at[i, 'md5']
        
        first_seen = df.at[i, 'first_seen']
        last_seen = df.at[i, 'last_seen']
        
        otx_url = 'https://otx.alienvault.com/api/v1/indicators/file/'+ str(sha256)
        
        try:
            for attempt in range(1, max_retries + 1):

                current_pulse = requests.get(otx_url,headers={"X-OTX-API-KEY":api_key})

                current_pulse = json.loads(current_pulse.content.decode("utf-8", "ignore"))

                tmp = pd.DataFrame.from_dict(pd.json_normalize(current_pulse)).reset_index(drop = True)
                
                if ('base_indicator.type' in tmp.columns) & ('base_indicator.indicator' in tmp.columns):
                    
                    found = False
                    
                    indicator_type = tmp.at[0, 'base_indicator.type']
                    indicator_base = tmp.at[0, 'base_indicator.indicator']
                    
                    if (indicator_type == 'FileHash-MD5') & ( indicator_base == md5 ):
                        found = True
                        
                    elif (indicator_type == 'FileHash-SHA1') & ( indicator_base == sha1 ): 
                        found = True
                        
                    if(found):
                        if ('pulse_info.count' in tmp.columns) & ('pulse_info.pulses' in tmp.columns):

                            if tmp.at[0,'pulse_info.count'] > 0:

                                tmp1 = pd.json_normalize(pd.DataFrame.from_dict(pd.json_normalize(tmp['pulse_info.pulses']).T).reset_index(drop = True)[0])

                                tmp1['sha256'] = sha256
                                tmp1['sha1'] = sha1
                                tmp1['md5'] = md5
                                
                                dataframe = pd.concat([dataframe, tmp1])
                                break
                                
                        else:
                            break
                                
                    else:
                        break
                        
                else:
                    break
                        
                
        except json.JSONDecodeError as e:
            print(f"JSON Decode Error on attempt {attempt}/{max_retries}: {e}")
            if attempt < max_retries:
                print(f"Retrying in {retry_delay} seconds...")
                time.sleep(retry_delay)
                
            else:
                print("Max retries reached. Unable to handle JSON Decode Error.")
                
        except (requests.exceptions.ReadTimeout, requests.exceptions.RequestException) as e:
            print(f"Request attempt {attempt} timed out. Retrying...")
            if attempt < max_retries:
                print(f"Retrying in {retry_delay} seconds...")
                time.sleep(retry_delay)

            else:
                print("Max retries reached. Unable to handle JSON Decode Error.")
                
    dataframe['scanner_date'] = datetime.now()   
            
    return dataframe.reset_index(drop = True)         

otx_pulses_df = create_otx_pulses_tdf(entries_df)
otx_pulses_df

otx_pulses_df.to_csv(r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\FileAnalysis\PostDataEnrichment\Step2\OTXPulses.csv')
