==> 23.2. NLP - Words - Twitter Posts.ipynb <==
import os
import time
import pandas as pd 
import requests
import urllib3
import json
import sys
import numpy as np
import pandas as pd
from pathlib import Path
from datetime import datetime, timedelta

pd.options.display.max_colwidth = 1000
pd.set_option('display.max_columns', None)

### Files Yara Rules

file_twitter_dir = r'C:\Users\leona\OneDrive\Desktop\Tesi\Pipeline\Datasets\Pipeline\SocialNetworksPosts\Twitter\TwitterPosts.csv'

df_twitter = pd.read_csv(file_twitter_dir, low_memory = False)
df_twitter = df_twitter.drop(columns=['Unnamed: 0', 'json'])
df_twitter

df_twitter['text'] = df_twitter['text'].fillna(' ')
df_twitter

df_twitter['date'] = pd.to_datetime(df_twitter['date']).dt.strftime('%Y-%m-%d')
df_twitter

### NLP sul Behaviours in ogni file

#### Example of NLP

!python -m spacy download it_core_news_sm


import spacy

nlp = spacy.load("it_core_news_sm")

frase = "Il gatto mangia il pesce."

doc = nlp(frase)

# Estrai sostantivi e verbi dalla frase
parole_di_interesse = [token.text for token in doc if token.pos_ in ['NOUN', 'VERB']]

# Stampa i risultati
print("Parole di interesse:", parole_di_interesse)

#### Get Words from descriptions

pip install spacy


!python -m spacy download en_core_web_sm

selected_rows = df_twitter.loc[(pd.to_datetime(df_twitter['date']).dt.month >= 6) & (pd.to_datetime(df_twitter['date']).dt.year == 2023) ].reset_index(drop =True)
selected_rows

from tqdm import tqdm

def create_words_df(df):
    nlp = spacy.load("en_core_web_sm")
    combined_texts = df['text'] 
    
    data_list = []
    
    for i, text in tqdm(enumerate(combined_texts), total=len(combined_texts), desc="Processing documents"):
        doc = nlp(text)
        words_interesting = [token.text for token in doc if token.pos_ in ['NOUN', 'VERB']]
        
        for word in words_interesting:
            word_type = 'noun' if nlp(word)[0].pos_ == 'NOUN' else 'verb'
            data_list.append({'word': word, 'type': word_type, 'index': i})
            
    dataframe = pd.DataFrame(data_list).reset_index()
    
    return dataframe


words_df = create_words_df(selected_rows)

words_df

words_df.to_csv(r'C:\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\SocialNetworksPosts\Twitter\PostsAnalysis.csv')

merged_words_df = pd.merge(selected_rows, words_df, left_index=True, right_on='index')
merged_words_df = merged_words_df.drop(columns=['index'])

merged_words_df.head()


merged_words_df.to_csv(r'C:\Users\\leona\\OneDrive\\Desktop\\Tesi\\Pipeline\\Datasets\\Pipeline\\SocialNetworksPosts\Twitter\PostsAnalysis.csv')
